requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
certifi==2024.8.30
click==8.1.7
colorama==0.4.6
fastapi==0.115.4
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httptools==0.6.4
httpx==0.27.2
idna==3.10
loguru==0.7.2
pillow==11.0.0
pydantic==2.9.2
pydantic_core==2.23.4
PyMySQL==1.1.1
pyotp==2.9.0
python-dotenv==1.0.1
python-telegram-bot==21.7
PyYAML==6.0.2
qrcode==8.0
redis==5.2.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.2
typing_extensions==4.12.2
uvicorn==0.32.0
watchfiles==0.24.0
websockets==14.0
win32-setctime==1.1.0
